Better Croatian translation
sect1_slogan: Updated to reflect the changes in the English version.
sect1_notavail: Added.
sect2_aboutbody: The verb stvoren (made) fits a lot better than napisan (written).
sect3_video_blurb: "Viđenje je vjerovanje" is a word-for-word translation of the English version and sounds weird. I believe "Slika vrijedi tisuću riječi" (A picture is worth a thousand words) is a better choice because it fits in the context and actually exists as a phrase in the Croatian language.
sect4_principles: Translated "IM".